refactor: split helper code to dedicated files
